requirements.txt
faster-whisper
feedparser
requests
tqdm
opencc-python-reimplemented